actually swarms now
lots of debug messages

code pretty hacked.

now on to prettyness, performance, balance and increasing swarmers

+++ src/main/scala/com/droidswarm/simulation/Vector.scala
@@ -1,6 +1,6 @@
 package com.droidswarm.simulation
 
-class Point(val x: Double, val y: Double) {
+class Point(val x: Float, val y: Float) {
 
   def + (operand: Vector): Vector = new Vector (x + operand.x, y + operand.y)
   def -> (operand: Point): Vector = new Vector (operand.x - x, operand.y - y)
@@ -9,16 +9,16 @@ class Point(val x: Double, val y: Double) {
   override def toString = x + "," + y
 }
 
-class Vector(x: Double, y: Double) extends Point(x,y) {
+class Vector(x: Float, y: Float) extends Point(x,y) {
 
-  def / (operand: Double): Vector = new Vector (x / operand, y / operand)
-  def * (operand: Double): Vector = new Vector (x * operand, y * operand)
+  def / (operand: Float): Vector = new Vector (x / operand, y / operand)
+  def * (operand: Float): Vector = new Vector (x * operand, y * operand)
   def * (operand: Int): Vector = new Vector (x * operand, y * operand)
 
   // Using java.lang.Math here not the scala.Math object as the scala version does not have
   // the handy dandy hypot method. It's not hard to calculate the hypot by hand but this way
   // is more terse
-  def magnitude: Double = java.lang.Math.hypot(x,y)
+  def magnitude: Float = (java.lang.Math.hypot(x,y)).toFloat
 
   def normalise: Vector = {
     this / magnitude
@@ -26,9 +26,9 @@ class Vector(x: Double, y: Double) extends Point(x,y) {
 
   def inverse = this * -1
 
-  def rotate(angle: Double): Vector = {
-    new Vector(x * Math.cos(angle) - y * Math.sin(angle),
-               x * Math.sin(angle) + y * Math.cos(angle))
+  def rotate(angle: Float): Vector = {
+    new Vector((x * Math.cos(angle) - y * Math.sin(angle)).toFloat,
+               (x * Math.sin(angle) + y * Math.cos(angle)).toFloat)
   }
 
 }
